Extract E1002 display functions to displays/spectra6.cpp
