+++ src/displays/spectra6.h
@@ -1,0 +1,5 @@
+#pragma once
+
+// Public API for the Spectra 6 (reTerminal E1002) panel
+bool spectra6_init_spi();
+bool spectra6_update();
